refactor: re-exported types to sdk and removed duplicates
