+++ near-sdk-core/src/lib.rs
@@ -2,14 +2,11 @@ pub(crate) mod di;
 
 pub mod json_types;
 
-mod state_init;
-pub use state_init::*;
+pub mod state_init;
 
-mod types;
-pub use crate::types::*;
+pub mod types;
 
-mod allowance;
-pub use crate::allowance::*;
+pub mod allowance;
 
 pub use base64;
 pub use borsh;
