+++ near-sdk-core/src/state_init.rs
@@ -4,7 +4,7 @@ use near_account_id::AccountId;
 use near_sdk_macros::near;
 use serde_with::base64::Base64;
 
-use crate::{GlobalContractId, di::EnvironmentBasedEnv};
+use crate::{di::EnvironmentBasedEnv, types::GlobalContractId};
 
 #[cfg_attr(feature = "arbitrary", derive(::arbitrary::Arbitrary))]
 #[near(inside_nearsdk, serializers = [
