+++ near-sdk-core/src/types/mod.rs
@@ -1,6 +1,3 @@
-mod vm_types;
-pub use self::vm_types::*;
-
 mod public_key;
 pub use self::public_key::{CurveType, PublicKey};
