+++ near-sdk/Cargo.toml
@@ -25,6 +25,7 @@ serde_json = { version = "1", features = ["preserve_order"] }
 serde_with = { version = "3", features = ["base64", "hex", "json"] }
 near-sdk-macros = { path = "../near-sdk-macros", version = "~5.24.1" }
 near-sys = { path = "../near-sys", version = "0.2.7" }
+near-sdk-core = { path = "../near-sdk-core/" }
 base64 = "0.22"
 borsh = { version = "1.0.0", features = ["derive"] }
 bs58 = "0.5"
@@ -79,12 +80,13 @@ insta = "1.39.0"
 [features]
 default = ["wee_alloc"]
 
-arbitrary = ["dep:arbitrary", "near-account-id/arbitrary"]
-expensive-debug = []
-deterministic-account-ids = [
-    "dep:hex",
-    "near-sys/deterministic-account-ids",
+arbitrary = [
+    "dep:arbitrary",
+    "near-account-id/arbitrary",
+    "near-sdk-core/arbitrary",
 ]
+expensive-debug = []
+deterministic-account-ids = ["dep:hex", "near-sys/deterministic-account-ids"]
 unstable = []
 legacy = []
 global-contracts = []
@@ -104,6 +106,7 @@ unit-testing = [
     "near-primitives",
     "near-crypto",
     "near-parameters",
+    "near-sdk-core/__near-sdk-unit-testing",
 ]
 non-contract-usage = []
